requirements.txt
kafka-python
pyspark
delta-spark
